requirements.txt
streamlit>=1.39.0
groq>=0.11.0
playwright>=1.48.0
python-dotenv>=1.0.0
youtube-transcript-api>=0.6.2
transformers>=4.45.0
torch>=2.5.0
sentencepiece>=0.2.0
psutil>=7.0.0
huggingface-hub>=0.26.0
